[CWS] split cilium using kernel check functions to gated file
